refactor(ui): clean up the field validation file
clean up the field validation file
